# base/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Học Tiếng Anh",
        "description": "Tài liệu API cho ứng dụng học tiếng Anh.",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/categories": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Tạo danh mục mới",
                "operationId": "3d09a8f4aea96aed370bc0120547a197",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "example": "Sách tiếng Anh"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tạo danh mục thành công"
                    },
                    "400": {
                        "description": "Tạo danh mục thất bại"
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Lấy chi tiết danh mục",
                "operationId": "b75c7d133a8bc87cc78ce8125324d822",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy thông tin danh mục thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "category_name": {
                                            "type": "string",
                                            "example": "Sách tiếng Anh"
                                        },
                                        "position": {
                                            "type": "integer",
                                            "example": 3
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục"
                    }
                }
            },
            "put": {
                "tags": [
                    "Category"
                ],
                "summary": "Cập nhật danh mục",
                "operationId": "8edd529bf8c5d2c024f479805defe8b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "example": "Sách luyện nghe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cập nhật thành công"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục"
                    }
                }
            }
        },
...
    },
...
}